Make StringConverters public, add package-info
