+++ src/main/java/module-info.java
@@ -11,5 +11,6 @@ module qupath.fx {
     exports qupath.fx.dialogs;
     exports qupath.fx.localization;
     exports qupath.fx.prefs;
+    exports qupath.fx.prefs.annotations;
     exports qupath.fx.utils;
 }
