+++ src/main/java/qupath/fx/controls/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Custom JavaFX controls for specific tasks.
+ */
+package qupath.fx.controls;
