+++ src/main/java/qupath/fx/dialogs/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Helper classes for generating dialogs, notifications and file or directory choosers.
+ */
+package qupath.fx.dialogs;
