+++ src/main/java/qupath/fx/prefs/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Create and manage JavaFX properties that can be made persistent using the Java Preferences API.
+ */
+package qupath.fx.prefs;
