+++ src/main/java/qupath/fx/utils/converters/EnumConverter.java
@@ -14,12 +14,16 @@
  * limitations under the License.
  */
 
-package qupath.fx.prefs.converters;
+package qupath.fx.utils.converters;
 
 import javafx.util.StringConverter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+/**
+ * General {@link StringConverter} for enums.
+ * @param <T> the enum type
+ */
 public class EnumConverter<T extends Enum> extends StringConverter<T> {
 
     private final static Logger logger = LoggerFactory.getLogger(EnumConverter.class);
