+++ src/main/java/qupath/fx/utils/converters/FileConverter.java
@@ -14,36 +14,32 @@
  * limitations under the License.
  */
 
-package qupath.fx.prefs.converters;
+package qupath.fx.utils.converters;
 
 import javafx.util.StringConverter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.Arrays;
-import java.util.Locale;
-import java.util.Objects;
+import java.io.File;
 
-public class LocaleConverter extends StringConverter<Locale> {
+/**
+ * General {@link StringConverter} for File objects, using their default string representation.
+ */
+public class FileConverter extends StringConverter<File> {
 
-    private final static Logger logger = LoggerFactory.getLogger(LocaleConverter.class);
+    private final static Logger logger = LoggerFactory.getLogger(FileConverter.class);
 
     @Override
-    public String toString(Locale locale) {
-        return locale == null ? null : locale.getDisplayName(Locale.US);
-//        return locale == null ? null : locale.toLanguageTag();
+    public String toString(File file) {
+        return file == null ? null : file.toString();
     }
 
     @Override
-    public Locale fromString(String string) {
+    public File fromString(String string) {
         if (string == null)
             return null;
         try {
-            return Arrays.stream(Locale.getAvailableLocales())
-                    .filter(l -> Objects.equals(l.getDisplayName(Locale.US), string))
-                    .findFirst().orElse(null);
-//            // Note the Locale javadocs state that toLanguageTag and forLanguageTag aren't guaranteed to round-trip!
-//            return Locale.forLanguageTag(string);
+            return new File(string);
         } catch (Exception e) {
             logger.error("Could not parse file from " + string, e);
             return null;
