+++ src/main/java/qupath/fx/utils/converters/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Implementations of a JavaFX {@link javafx.util.StringConverter} for common types.
+ */
+package qupath.fx.utils.converters;
