+++ src/main/java/qupath/fx/utils/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Utility functions to help with JavaFX.
+ */
+package qupath.fx.utils;
