+++ src/test/java/qupath/fx/prefs/converters/TestUriConverter.java
@@ -18,6 +18,7 @@ package qupath.fx.prefs.converters;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
+import qupath.fx.utils.converters.UriConverter;
 
 import java.net.URI;
 import java.nio.file.Paths;
